changed variable types int/ decimal to float64

+++ go/src/backtesting/utils.go
@@ -9,9 +9,9 @@ type ticker func(int) decimal.Decimal
 
 // struct for state of trader
 type state_t struct {
-  funds int
-  assets int
-  inventory [][]int
+  funds float64
+  assets float64
+  inventory [][]float64
   historicalData [][][]string
   currentDay int
   metrics struct  {
